requirements.txt
pathtools==0.1.2
PyCocoa==20.1.2
pyobjc-core==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-FSEvents==6.1
PyQt5==5.14.1
PyQt5-sip==12.7.0
python-vlc==3.0.7110
watchdog==0.10.0
